Move milestone types into 'types' module
