+++ bee-inx/src/lib.rs
@@ -14,8 +14,7 @@ pub mod utxo;
 pub use self::{block::*, error::Error, milestone::*, node::*, raw::*, request::*, utxo::*};
 
 pub mod inx {
-    pub use ::inx::proto::*;
-    pub use ::inx::proto::block_metadata::*;
+    pub use ::inx::proto::{block_metadata::*, *};
 }
 
 #[macro_export]
